refactor(log): move create-log.dto to domain/log/dtos directory
Restructure DTO location to better organize domain files by feature/module
