+++ src/infrastructure/services/logger.service.ts
@@ -1,6 +1,6 @@
 import { LogDatasource } from "../datasources/log.datasource";
 import { TypeGuardsUtils } from "@/utils/type-guards.utils";
-import { CreateLogDto } from "@/domain/dtos/create-log.dto";
+import { CreateLogDto } from "@/domain/log/dtos/create-log.dto";
 import { CreateLog } from "@/domain/use-cases/create-log";
 import { ILogData, PROD_LOG_LEVELS } from "@/domain/interfaces/log.interfaces";
 import envs from "@/config/envs";
