+++ src/infrastructure/test-infrastructure.ts
@@ -1,7 +1,7 @@
 import { CreateLogValidator } from "./validators/create-log.validator";
 import { LogDatasource } from "./datasources/log.datasource";
 import { DatabaseClient } from "./config/database.client";
-import { CreateLogDto } from "@/domain/dtos/create-log.dto";
+import { CreateLogDto } from "@/domain/log/dtos/create-log.dto";
 
 /**
  * Test script for infrastructure layer components
