+++ src/infrastructure/validators/create-log.validator.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { BadRequestError } from "@/domain/errors/bad-request-error";
-import { CreateLogDto } from "@/domain/dtos/create-log.dto";
+import { CreateLogDto } from "@/domain/log/dtos/create-log.dto";
 import { LOG_LEVELS, type LogLevel } from "@/domain/interfaces/log.interfaces";
 import type { TRawJson } from "@/domain/interfaces/general.interfaces";
 import { processValidationError } from "./utils/processError.validator";
